FIX issue #298: removed stale reference
self._power_service is initialized upon startup, but never refreshed. Subsequent updates on the state of the device (power socket) will not be propagated to self._power_service. is_on(...) now fetches the fresh state from the coordinator.

Review is strongly recommended. I have minimal Python knowledge and It took me quite some trial and error and some discussion with an AI to get it working. I am actually convinced that there is a cleaner and neater solution to the issue.

Handle this change proposition at your own discretion.